[YA-1931] fix archive creation (#567)
* [YA-1931] fix company archive export

+++ app/services/export/create_archive_zip.rb
@@ -29,6 +29,7 @@ module Export
         FileUtils.mkdir_p(employee_dir_path)
 
         employee.files.each do |employee_file|
+          next unless File.exist?(employee_file.file.path)
           FileUtils.cp(employee_file.file.path, employee_dir_path)
         end
       end
